update get method for problem and instance_data
