+++ .streamlit/config.toml
@@ -1,0 +1,2 @@
+[theme]
+backgroundColor="#FFFFFF"
