+++ jijbench/__init__.py
@@ -6,8 +6,7 @@ import jijbench.functions as functions
 import jijbench.node as node
 from jijbench.benchmark.benchmark import Benchmark, construct_benchmark_for
 from jijbench.containers.containers import Artifact, Record, Table
-from jijbench.datasets.instance_data import get_instance_data
-from jijbench.datasets.problem import get_problem
+from jijbench.datasets.model import get_instance_data, get_models, get_problem
 from jijbench.elements.array import Array
 from jijbench.elements.base import Callable, Number, String
 from jijbench.elements.date import Date
@@ -16,13 +15,14 @@ from jijbench.evaluation.evaluation import Evaluation
 from jijbench.experiment.experiment import Experiment
 from jijbench.io.io import load, save
 from jijbench.solver.base import Parameter, Response, Solver
-from jijbench.solver.jijzept import InstanceData, SampleSet, UserDefinedModel
+from jijbench.solver.jijzept import InstanceData, Model, SampleSet
 
 __all__ = [
     "construct_benchmark_for",
     "functions",
     "node",
     "get_instance_data",
+    "get_models",
     "get_problem",
     "load",
     "save",
@@ -35,7 +35,7 @@ __all__ = [
     "Experiment",
     "ID",
     "InstanceData",
-    "UserDefinedModel",
+    "Model",
     "Number",
     "Parameter",
     "Record",
