+++ jijbench/consts/default.py
@@ -1,0 +1,10 @@
+import pathlib
+
+DEFAULT_RESULT_DIR = pathlib.Path("./.jb_results")
+DEFAULT_PROBLEM_NAMES = [
+    "bin-packing",
+    "knapsack",
+    "nurse-scheduling",
+    "travelling-salesman",
+    "travelling-salesman-with-time-windows",
+]
