--- jijbench/consts/path.py
@@ -1,3 +1,0 @@
-import pathlib
-
-DEFAULT_RESULT_DIR = pathlib.Path("./.jb_results")
