+++ jijbench/datasets/problems/__init__.py
@@ -1,0 +1,15 @@
+from jijbench.datasets.problems.bin_packing import bin_packing
+from jijbench.datasets.problems.knapsack import knapsack
+from jijbench.datasets.problems.nurse_scheduling import nurse_scheduling
+from jijbench.datasets.problems.travelling_salesman import travelling_salesman
+from jijbench.datasets.problems.travelling_salesman_with_time_windows import (
+    travelling_salesman_with_time_windows,
+)
+
+__all__ = [
+    "bin_packing",
+    "knapsack",
+    "nurse_scheduling",
+    "travelling_salesman",
+    "travelling_salesman_with_time_windows",
+]
