+++ pyproject.toml
@@ -45,6 +45,11 @@ matplotlib = "^3.7.1"
 nbformat = "^5.8.0"
 networkx = "^3.0"
 dill = "^0.3.6"
+streamlit = "^1.20.0"
+streamlit-aggrid = "^0.3.4"
+streamlit-ace = "^0.1.1"
+streamlit-elements = "^0.1.0"
+streamlit-tree-select = "^0.0.5"
 jijmodeling = "^0.10.21"
 jijzept = "^1.14.0"
